Web View fix
fixed init of WVC

+++ NewsReader/Presentation/Favorites Table View/FavoritesTableViewController.swift
@@ -11,49 +11,38 @@ import CoreData
 
 class FavoritesTableViewController: UITableViewController, NSFetchedResultsControllerDelegate {
     var managedContext: NSManagedObjectContext!
-    
     var fetchedResultsController: NSFetchedResultsController<NSFetchRequestResult>!
-    
     let favoritesCellIdentifier = "FavoritesCell"
-    
     let favoritesExitSegueIdentifier = "FavoritesExitSegue"
 
     override func viewDidLoad() {
         super.viewDidLoad()
         self.managedContext = CoreDataStack.instance.context
         let fetchRequest = NSFetchRequest<Favorite>(entityName: "Favorite")
-        
         let sortDescriptor = NSSortDescriptor(key: "name", ascending: true)
         fetchRequest.sortDescriptors = [sortDescriptor]
-        
-        self.fetchedResultsController = NSFetchedResultsController(fetchRequest: fetchRequest, managedObjectContext: self.managedContext, sectionNameKeyPath: nil, cacheName: nil) as? NSFetchedResultsController<NSFetchRequestResult>
+        self.fetchedResultsController = NSFetchedResultsController(fetchRequest: fetchRequest,
+                                                                   managedObjectContext: self.managedContext,
+                                                                   sectionNameKeyPath: nil,
+                                                                   cacheName: nil) as? NSFetchedResultsController<NSFetchRequestResult>
         self.fetchedResultsController.delegate = self
-        
         do {
             try self.fetchedResultsController.performFetch()
         } catch let error as NSError {
             print("Error:\(error.localizedDescription)")
         }
     }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-    }
-    
-    // MARK: - Button actions
-    
     @IBAction func addFavoriteAction(_ sender: Any) {
         let alert = UIAlertController(title: "Favorites", message: "Add new RSS Channel", preferredStyle: .alert)
-        
-        let addFavoriteAction = UIAlertAction(title: "Add", style: .default, handler: { (action: UIAlertAction!) in
+        let addFavoriteAction = UIAlertAction(title: "Add", style: .default, handler: { (_: UIAlertAction!) in
             let nameTextField = alert.textFields![0]
             let linkTextField = alert.textFields![1]
-            
-            let favorite = NSEntityDescription.insertNewObject(forEntityName: "Favorite", into: self.managedContext) as! Favorite
-            
+            guard let favorite = NSEntityDescription.insertNewObject(forEntityName: "Favorite",
+                                                                     into: self.managedContext) as? Favorite else {
+                                                                        return
+            }
             favorite.name = nameTextField.text
             favorite.link = linkTextField.text
-            
             do {
                 try self.managedContext.save()
             } catch let error as NSError {
@@ -61,14 +50,14 @@ class FavoritesTableViewController: UITableViewController, NSFetchedResultsContr
             }
         })
         addFavoriteAction.isEnabled = false
-        
         alert.addTextField { (textField: UITextField!) in
             textField.placeholder = "Name"
         }
         alert.addTextField { (textField: UITextField!) in
             textField.placeholder = "Link"
-            
-            NotificationCenter.default.addObserver(forName: UITextField.textDidChangeNotification, object: textField, queue: OperationQueue.main) { (notification) in
+            _ = NotificationCenter.default.addObserver(forName: UITextField.textDidChangeNotification,
+                                                   object: textField,
+                                                   queue: OperationQueue.main) { (_) in
                 if let link = textField.text {
                     if link.hasSuffix(".xml") || link.hasSuffix(".rss") {
                         addFavoriteAction.isEnabled = true
@@ -76,15 +65,10 @@ class FavoritesTableViewController: UITableViewController, NSFetchedResultsContr
                 }
             }
         }
-        
         alert.addAction(addFavoriteAction)
         alert.addAction(UIAlertAction(title: "Cancel", style: .default, handler: nil))
-        
         present(alert, animated: true, completion: nil)
     }
-    
-    // MARK: - Table view data source
-
     override func numberOfSections(in tableView: UITableView) -> Int {
         return fetchedResultsController.sections!.count
     }
@@ -97,8 +81,7 @@ class FavoritesTableViewController: UITableViewController, NSFetchedResultsContr
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: self.favoritesCellIdentifier, for: indexPath)
 
-        let favorite = fetchedResultsController.object(at: indexPath) as! Favorite
-        
+        guard let favorite = fetchedResultsController.object(at: indexPath) as? Favorite else {return UITableViewCell()}
         cell.textLabel?.text = favorite.name
         cell.detailTextLabel?.text = favorite.link
 
@@ -107,11 +90,11 @@ class FavoritesTableViewController: UITableViewController, NSFetchedResultsContr
     override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
         return true
     }
-    
-    override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
+    override func tableView(_ tableView: UITableView,
+                            commit editingStyle: UITableViewCell.EditingStyle,
+                            forRowAt indexPath: IndexPath) {
         if editingStyle == .delete {
-            let favorite = fetchedResultsController.object(at: indexPath) as! Favorite
-            
+            guard let favorite = fetchedResultsController.object(at: indexPath) as? Favorite else {return}
             self.managedContext.delete(favorite)
             do {
                 try self.managedContext.save()
@@ -120,15 +103,14 @@ class FavoritesTableViewController: UITableViewController, NSFetchedResultsContr
             }
         }
     }
-    
-    // MARK: - NSFetchedResultsControllerDelegate
-    
-    func controllerWillChangeContent(_ controller:
-        NSFetchedResultsController<NSFetchRequestResult>) {
+    func controllerWillChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
             self.tableView.beginUpdates()
     }
-    
-    func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>, didChange anObject: Any, at indexPath: IndexPath?, for type: NSFetchedResultsChangeType, newIndexPath: IndexPath?) {
+    func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>,
+                    didChange anObject: Any,
+                    at indexPath: IndexPath?,
+                    for type: NSFetchedResultsChangeType,
+                    newIndexPath: IndexPath?) {
         if type == .insert {
             if let newIndex = newIndexPath {
             self.tableView.insertRows(at: [newIndex], with: .automatic)
@@ -144,7 +126,6 @@ class FavoritesTableViewController: UITableViewController, NSFetchedResultsContr
             }
         }
     }
-    
     func controllerDidChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
         tableView.endUpdates()
     }
@@ -155,7 +136,7 @@ class FavoritesTableViewController: UITableViewController, NSFetchedResultsContr
         if segue.identifier == favoritesExitSegueIdentifier {
             if let destination = segue.destination as? NewsTableViewController {
                 if let indexPath = self.tableView.indexPathForSelectedRow {
-                    let favorite = fetchedResultsController.object(at: indexPath) as! Favorite
+                    guard let favorite = fetchedResultsController.object(at: indexPath) as? Favorite else {return}
                     guard let link = favorite.link else {
                         return
                     }
